refactor: extract useSpellInput hook from App.tsx
